Moved PotentialCoins to a different file
